refactor(Template): rename Template.stories.tsx to ElmTemplate.stories.tsx and update component import
